integrating external search flow (WIP)
