+++ pyproject.toml
@@ -44,6 +44,7 @@ rank-bm25 = "^0.2.2"
 psycopg = {extras = ["binary"], version = "^3.2.3"}
 pydantic = "^2.10.5"
 pyyaml = "^6.0.2"
+tavily-python = "^0.5.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.7.1"
@@ -54,6 +55,8 @@ isort = "^5.13.2"
 pandas-stubs = "^2.2.3.241009"
 types-requests = "^2.32.0.20241016"
 types-pyyaml = "^6.0.12.20241230"
+datasets = "^3.2.0"
+ragas = "^0.2.11"
 
 [[tool.poetry.source]]
 name = "PyPI"
